OnlineExam/empDash_Goalsheet_OnlineExam/requirements20181219.txt
APScheduler==3.5.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
et-xmlfile==1.0.1
Flask==0.12.2
Flask-APScheduler==1.8.0
Flask-Cache==0.13.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.2
openpyxl==2.5.1
pandas==0.22.0
phonenumbers==8.9.13
PyMySQL==0.8.0
python-dateutil==2.7.1
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
tzlocal==1.5.1
urllib3==1.22
waitress==1.1.0
Werkzeug==0.14.1
WTForms==2.1
xlrd==1.1.0
